Move getUserProfile to users file
