+++ client/src/js/actions/users.test.js
@@ -1,6 +1,6 @@
 import axios from 'axios';
 import * as actions from './users';
-import { REGISTER_USER, LOGIN_USER } from "./types";
+import { GET_USER, REGISTER_USER, LOGIN_USER } from "./types";
 
 jest.mock('axios');
 jest.mock('../utils/setAuthToken');
@@ -86,5 +86,47 @@ describe('actions', () => {
 		const dispatch = jest.fn();
 		actions.logoutUser()(dispatch);
 		expect(dispatch).toHaveBeenCalledWith({"payload": {}, "type": "SET_CURRENT_USER"});
+	});
+
+	describe('getUserProfile', () => {
+		const resp = {
+			data: {
+			rating: 0,
+			name: "test name",
+			email: "test1@test.com",
+			_id: "5dbb0cc7edeea12284986dcc"
+			}
+		}
+		test('Successful getUserProfile', async () => {
+			const dispatch = jest.fn();
+			let mock = axios.get.mockResolvedValue(resp)
+			await actions.getUserProfile(resp.data._id)(dispatch)
+			expect(mock).toHaveBeenCalledWith('/api/users/'+resp.data._id) //asserting what im testing in line 87
+		});
+
+		test('Successful dispatched', async () => {
+			const dispatch = jest.fn();
+			axios.get.mockResolvedValue(resp)
+			await actions.getUserProfile(resp)(dispatch)
+			expect(dispatch).toHaveBeenCalledWith({
+				type: GET_USER.FINISHED,
+				payload: resp.data
+			});
+		});
+
+		test('error dispatched', async () => {
+			const dispatch = jest.fn();
+			const mock_error = {
+				data: 'test error',
+				type: LOGIN_USER.ERROR,
+	      	}
+			//axios.get.mockRejectedValue(resp);
+			axios.get.mockRejectedValueOnce(mock_error);
+		  	await actions.getUserProfile(mock_error)(dispatch);
+		  	expect(dispatch).toHaveBeenCalledWith({
+		    	type: GET_USER.ERROR,
+		    	payload: mock_error
+		  	});
+		});
 	});
 });
