+++ client/src/js/components/Profile/Profile.js
@@ -1,7 +1,8 @@
 import React, { Component } from "react";
 import ProfileCard from './ProfileCard'
 import { connect } from "react-redux";
-import { getUserProfile, fetchProducts } from "../../actions";
+import { fetchProducts } from "../../actions";
+import { getUserProfile } from "../../actions/users";
 
 export class UserProfile extends Component {
     constructor(props) {
